Refactor regression into sr subpackage
Reorganize the regression package by moving the symbolic-regression modules into regression/sr and adding a regression.ols package. Updated regression/__init__ to expose ols and sr, created sr/__init__.py, and adjusted imports across the codebase (core, tests, and docs) to use the new regression.sr paths. Also fixed relative imports (e.g. sr_interface, solved_model) to match the new layout. This change separates symbolic regression implementation from other regressors and cleans up the package structure.
