+++ SymbolicDSGE/regression/__init__.py
@@ -1,13 +1,6 @@
-from .model_parametrizer import ModelParametrizer
-from .model_defaults import make_operator_general, CustomOp, PySRParams
-from .config import TemplateConfig
-from .symbolic_regression import SymbolicRegressor
+from . import ols, sr
 
 __all__ = [
-    "ModelParametrizer",
-    "make_operator_general",
-    "CustomOp",
-    "PySRParams",
-    "TemplateConfig",
-    "SymbolicRegressor",
+    "ols",
+    "sr",
 ]
