+++ SymbolicDSGE/regression/sr/built_in_op_container.py
