+++ SymbolicDSGE/regression/sr/fit_result.py
