+++ SymbolicDSGE/regression/sr/template_factory.py
