+++ tests/core/test_solved_model_fit_kf_contract.py
@@ -6,9 +6,9 @@ import pytest
 
 import SymbolicDSGE.core.solved_model as solved_model_module
 from SymbolicDSGE.core.solved_model import SolvedModel
-from SymbolicDSGE.regression.config import TemplateConfig
-from SymbolicDSGE.regression.model_defaults import PySRParams
-from SymbolicDSGE.regression.model_parametrizer import ModelParametrizer
+from SymbolicDSGE.regression.sr.config import TemplateConfig
+from SymbolicDSGE.regression.sr.model_defaults import PySRParams
+from SymbolicDSGE.regression.sr.model_parametrizer import ModelParametrizer
 
 
 def _make_solved_model() -> SolvedModel:
