+++ tests/regression/test_sr_backend_contract.py
@@ -10,10 +10,10 @@ import pytest
 import sympy as sp
 from pysr import ExpressionSpec, TemplateExpressionSpec
 
-from SymbolicDSGE.regression.config import TemplateConfig
-from SymbolicDSGE.regression.model_defaults import PySRParams
-from SymbolicDSGE.regression.model_parametrizer import ModelParametrizer
-from SymbolicDSGE.regression.sr_backend import SymbolicRegressorBackend
+from SymbolicDSGE.regression.sr.config import TemplateConfig
+from SymbolicDSGE.regression.sr.model_defaults import PySRParams
+from SymbolicDSGE.regression.sr.model_parametrizer import ModelParametrizer
+from SymbolicDSGE.regression.sr.sr_backend import SymbolicRegressorBackend
 
 
 @dataclass
